lib: Fixed Visual Studio warning: implicit casts between integer types.
The change for ThrowJVMError is probably an actual bug fix,
rather than just stopping the compiler complaining.

(Edited by Chris Frey to only include the ThrowJVMError fix.)

+++ src/m_jvmdebug.cc
@@ -254,7 +254,7 @@ void JVMDebug::Detach()
 }
 
 
-void JVMDebug::ThrowJVMError(const std::string &msg, uint8_t cmd)
+void JVMDebug::ThrowJVMError(const std::string &msg, uint16_t cmd)
 {
 	std::ostringstream oss;
 	oss << msg << ": " << _("unexpected packet command code: ")
